MusicPuzzle & TelephonePuzzle Finish

+++ Assets/_Scripts/StunDetector.cs
@@ -7,8 +7,8 @@ public class StunDetector : MonoBehaviour {
 
 	IEnumerator OnTriggerEnter2D(Collider2D c)
 	{
-		float o_speed = c.gameObject.GetComponent<playerController> ().get_speed();
-		c.gameObject.GetComponent<playerController> ().set_speed (0);
+		float o_speed = (float) c.gameObject.GetComponent<playerController>().get_speed();
+		c.gameObject.GetComponent<playerController>().set_speed (0);
 		yield return new WaitForSeconds (stunTime);
 		c.gameObject.GetComponent<playerController> ().set_speed (o_speed);
 		Destroy (gameObject);
